commandline tool fixed
